fix interrupt\halt timing, passing 6 more tests
When on halt, there is a 2 T-cycles delay beetween fetching enabled
flagged interrupts and checking if there is one. This was the real
solution for the failing in the vblank_stat test that I tried to fix in
the last commit, as I have commented there.

I found this issue by comparing withh the SameBoy source code, I don't know
if this timing is documented else where.