fix bench by always running the same section of code
In the previous commit the bench was implement by diving the total
run time in n sections, and compute the mean and error from that.

That don't make much sense, each section run have diferent code, and I
was also including that variance in the error. I fix this now by always
reseting the gameboy before it run.

fix compile error