don't update ppu and sound on load/save state
This was causing a bug where clock count was loaded from the save state
first, and the update would run for millions of cycles, accumulating too
many samples in the SoundController output.

Did not cause any other bugs, because only the SoundController output
was not loaded from the save state.