read the header of each rom for the rom list
Need to fix the rom loading blocking the main thread (which is also UI
thread).

The rom list is being reload evertime the gui is rebuild, which is bad
for Android.

On Android reading files is really slow (>30ms per file), need a
solutions for that. (A list with 544 files takes 20 seconds to display).