change icon resolution based on scale_factor
The icons in giui does not scale with the dpi aware scale factor of the
UI, to preserve pixel perfectness. This redraws the icons in svg, and
render them in resolutions: the current one, and a 2 times bigger.

In this process I also put the gamepad and icons in the same texture.

Also remove the 'white.png' texture, and instead generate them at
runtime (which is faster, and uses less memory).