fix STOP instructions
The interpreter was not consuming the second byte of the STOP
instructions (which is ignored), making the pc counter disalign.

Also, when the CPU enters in Stopped state, it was not advancing the
clock, keeping the emulator in a infinite loop. Fixed it by mirroring
the behavior of the Halt state, but I didn't reserch what the proper
behaviour should be.