add a struct for the save state file header
This header contains the magic constant, the version (increased by this
commit) and a timestamp.

This patch was originally created to include a timestamp which would be
used by the functionality in the last commit. But only after
implementing this patch that I noticed that this would not help with
that functionality. But this will be useful in the future.