use deflate compression for Timeline savestates
This compress the savestates in the Timeline to allow more frames in the
Rewind. This achieves a compress ratio of about 20% to 30% of the
original size.

Reuses flate2 crate from the dependencies of image, so should not
increase binary size by much (not verified).

Not sure if it is worth it, has quite a big performance impact, but not
enough to lower it below 60 fps. Maybe implementing delta compression or
decreasing the savestate frequency later on will make the compression be
more tolerable.