fix draw_scan_line sprite rendering
The sprites were checking against the color of the lcd, instead of the
color of background, before applying the pallete.

Now the backgrond palleting and the sprites drawing are defered. The
sprite pixels are written to unused bits of screen bytes first.

There must be a better way of doing, but it is now more accurate, at
least.