disassembler: remove RAM tracing and fix bank tracing
Previously, tracing were assuming that the lower_bank could always be
computed from the upper_bank, which is not true. This changes fix that
by first returning the current pair of mapped banks in
Cartridge::curr_bank, and propagating the change through the codebase.

RAM tracing were removed in the middle of this, because it were not
being used, and the implementation was duplicating too much logic.