gui: update frame only once per redraw
The UpdateFrame event is triggered on every vblank, and is used to
update the debug view of the game.

But the game triggers vblank multiple times per frame, faster than
they are handle, frezzing the gui. To fix that, the UpdateFrame
just assign a bool, that is checked only once each redraw.