emulator: don't rely on past values of Instant
Previously, I used a Instant keeping the theorical time that the gameboy
was turned on to calculate the ammount of clocks in order to emulate the
gameboy in real time. Every time I fast-forward the emulate I recomputed
the start Instant to `Instant::now() - clock_to_duration(gb.clock_count)`,
getting a value of Instant in the past.

But it is not guarranteed that Instant is able to represent times in the
past and may panic on underflow (See https://github.com/sebcrozet/instant/issues/45).

This is fixed by keeping `last_start_time: Instant` as the time that
the gameboy started running at the clock count `last_start_clock`.