jit: make next_interrupt checks more refined
The last commit has increase the length of each block, holping for
decreasing the time spend in querying for blocks. But that change make
the emulador fallback to the interpreter more often, because now it is
more likely that the next interrupt will happens inside the next block.

This commit address that by changing the check that verifies if the next
interrupt happens before the block end, to check if it will happen until
the next check.