intepreter: skip ahead to next interrupt, when halted
This may ahead too much ahead, exceeding too much the timeout set  in
the main loop. I set a limit of 1 frame, so it will not skip a entire
frame at least, but it may still block the main loop from processing
input, increasing input delay.

I the above is theorical, didn't test any of that.