jit: keep running the interpreter if it is still need
The interpreter loop were ending when halted or doing any jump. But in
the next loop, because of the halt it would fallback again to the
interpreter, same thing if it was jumping outside ROM.

Avoid it by refining the loop exit conditions. Also added a timeout in
cause it stay for to long there.