draw_scan_line: simplify loop when there are 0 sprites
While drawing sprites, information is bit packet in a byte, that later
need to be unpack and check for construction the final pixel value.

But in the case where sprite_buffer_len is 0, this can be greatly
simplified.

Decrease the time spend in draw_scan_line by more than 3%, when benching
the first 6000 frames of Kirby's Dream Land.