core: test and fix that `reset_after_boot` is correct
When a boot is not provided, the GameBoy call `reset_after_boot` that
reset the GameBoy to the state where it should have been after finishing
running the boot rom.

Adds a test that checks if this is really happening (and fix it, because
it wasn't).

To run this test, a dump of the GameBoy boot rom must be provided.