core: add wave trace generation
VCD (value change dump) is a format used by EDA tools (like verilog
simulators) that shows the change of signals over time. This will dump
the change of every register and internal state of the emulator over
time.

This will create a file in `wave_trace/trace.vcd`, which can be opened
on GTKWave or Surfer for example

I have the idea of implementing it while playing with msinger/dmg-sim
simulation of DMG CPU B in verilog. This will make it easier to compare
my emulator with a simulation of the original hardware.