core: try to read cartridge even if invalid
Will now deduce the size of the rom and ram if they are invalid, and
ignore things like the header checksum. In does cases `Cartridge::new`
will return a `Err((erro, Some(cartridge)))`. The ui frontned is just
printing the error message and using the cartridge, but ideally it would
prompt the user for confirmation.